+++ src/LaraPlans/Events/SubscriptionPlanChanged.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Gerardojbaez\LaraPlans\Events;
+
+use Gerardojbaez\LaraPlans\Models\PlanSubscription;
+use Illuminate\Queue\SerializesModels;
+
+class SubscriptionPlanChanged
+{
+    use SerializesModels;
+
+    /**
+     * @var \LaraPlans\Models\PlanSubscription
+     */
+    public $subscription;
+
+    /**
+     * Create a new event instance.
+     *
+     * @param  \LaraPlans\Models\PlanSubscription  $subscription
+     * @return void
+     */
+    public function __construct(PlanSubscription $subscription)
+    {
+        $this->subscription = $subscription;
+    }
+}